// src/api/models/GeneralPowers.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PowerRow } from './PowerRow';
import {
    PowerRowFromJSON,
    PowerRowFromJSONTyped,
    PowerRowToJSON,
    PowerRowToJSONTyped,
} from './PowerRow';
import type { General } from './General';
import {
    GeneralFromJSON,
    GeneralFromJSONTyped,
    GeneralToJSON,
    GeneralToJSONTyped,
} from './General';

/**
 * One player's power habits on one general.
 * @export
 * @interface GeneralPowers
 */
export interface GeneralPowers {
    /**
     * 
     * @type {General}
     * @memberof GeneralPowers
     */
    general: General;
    /**
     * 
     * @type {number}
     * @memberof GeneralPowers
     */
    games: number;
    /**
     * 
     * @type {number}
     * @memberof GeneralPowers
     */
    minutes: number;
    /**
     * 
     * @type {number}
     * @memberof GeneralPowers
     */
    groupGames: number;
    /**
     * 
     * @type {number}
     * @memberof GeneralPowers
     */
    reconPerMinute: number;
    /**
     * 
     * @type {number}
     * @memberof GeneralPowers
     */
    groupReconPerMinute: number;
    /**
     * 
     * @type {Array<PowerRow>}
     * @memberof GeneralPowers
     */
    rows: Array<PowerRow>;
}



/**
 * Check if a given object implements the GeneralPowers interface.
 */
export function instanceOfGeneralPowers(value: object): value is GeneralPowers {
    if (!('general' in value) || value['general'] === undefined) return false;
    if (!('games' in value) || value['games'] === undefined) return false;
    if (!('minutes' in value) || value['minutes'] === undefined) return false;
    if (!('groupGames' in value) || value['groupGames'] === undefined) return false;
    if (!('reconPerMinute' in value) || value['reconPerMinute'] === undefined) return false;
    if (!('groupReconPerMinute' in value) || value['groupReconPerMinute'] === undefined) return false;
    if (!('rows' in value) || value['rows'] === undefined) return false;
    return true;
}

export function GeneralPowersFromJSON(json: any): GeneralPowers {
    return GeneralPowersFromJSONTyped(json, false);
}

export function GeneralPowersFromJSONTyped(json: any, ignoreDiscriminator: boolean): GeneralPowers {
    if (json == null) {
        return json;
    }
    return {
        
        'general': GeneralFromJSON(json['general']),
        'games': json['games'],
        'minutes': json['minutes'],
        'groupGames': json['groupGames'],
        'reconPerMinute': json['reconPerMinute'],
        'groupReconPerMinute': json['groupReconPerMinute'],
        'rows': ((json['rows'] as Array<any>).map(PowerRowFromJSON)),
    };
}

export function GeneralPowersToJSON(json: any): GeneralPowers {
    return GeneralPowersToJSONTyped(json, false);
}

export function GeneralPowersToJSONTyped(value?: GeneralPowers | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'general': GeneralToJSON(value['general']),
        'games': value['games'],
        'minutes': value['minutes'],
        'groupGames': value['groupGames'],
        'reconPerMinute': value['reconPerMinute'],
        'groupReconPerMinute': value['groupReconPerMinute'],
        'rows': ((value['rows'] as Array<any>).map(PowerRowToJSON)),
    };
}