// src/api/models/PowerRow.ts
export interface PowerRow {
    /**
     * 
     * @type {string}
     * @memberof PowerRow
     */
    power: string;
    /**
     * 
     * @type {boolean}
     * @memberof PowerRow
     */
    purchasable: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PowerRow
     */
    unlocksUnit: boolean;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    gamesPicked: number;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    pickRate: number;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    groupPickRate: number;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    avgPickMinute?: number | null;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    groupAvgPickMinute?: number | null;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    avgLevels: number;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    groupAvgLevels: number;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    uses: number;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    usesPerMinute: number;
    /**
     * 
     * @type {number}
     * @memberof PowerRow
     */
    groupUsesPerMinute: number;
}