// src/api/models/PowerStats.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PlayerPowerProfile } from './PlayerPowerProfile';
import {
    PlayerPowerProfileFromJSON,
    PlayerPowerProfileFromJSONTyped,
    PlayerPowerProfileToJSON,
    PlayerPowerProfileToJSONTyped,
} from './PlayerPowerProfile';

/**
 * The powers page payload: who can be picked, and the picked player.
 * @export
 * @interface PowerStats
 */
export interface PowerStats {
    /**
     * 
     * @type {Array<string>}
     * @memberof PowerStats
     */
    players: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof PowerStats
     */
    matches: number;
    /**
     * 
     * @type {PlayerPowerProfile}
     * @memberof PowerStats
     */
    profile?: PlayerPowerProfile | null;
}

/**
 * Check if a given object implements the PowerStats interface.
 */
export function instanceOfPowerStats(value: object): value is PowerStats {
    if (!('players' in value) || value['players'] === undefined) return false;
    if (!('matches' in value) || value['matches'] === undefined) return false;
    return true;
}

export function PowerStatsFromJSON(json: any): PowerStats {
    return PowerStatsFromJSONTyped(json, false);
}

export function PowerStatsFromJSONTyped(json: any, ignoreDiscriminator: boolean): PowerStats {
    if (json == null) {
        return json;
    }
    return {
        
        'players': json['players'],
        'matches': json['matches'],
        'profile': json['profile'] == null ? undefined : PlayerPowerProfileFromJSON(json['profile']),
    };
}

export function PowerStatsToJSON(json: any): PowerStats {
    return PowerStatsToJSONTyped(json, false);
}

export function PowerStatsToJSONTyped(value?: PowerStats | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'players': value['players'],
        'matches': value['matches'],
        'profile': PlayerPowerProfileToJSON(value['profile']),
    };
}